Don't need to convert to int from Integer for AppdefEntityID

+++ src/org/hyperic/hq/autoinventory/AIHistory.java
@@ -6,7 +6,7 @@
  * normal use of the program, and does *not* fall under the heading of 
  * "derived work". 
  *  
- * Copyright (C) [2004, 2005, 2006], Hyperic, Inc. 
+ * Copyright (C) [2004-2008], Hyperic, Inc. 
  * This file is part of HQ.         
  *  
  * HQ is free software; you can redistribute it and/or modify 
@@ -239,7 +239,7 @@ public class AIHistory extends PersistedObject
     public String getEntityName()
     {
         AppdefEntityID id = new AppdefEntityID(getEntityType().intValue(),
-                                               getEntityId().intValue());
+                                               getEntityId());
         return id.toString();
     }
